chore: move getGossipPeers to separate file
