+++ src/heartbeat.js
@@ -1,7 +1,8 @@
 'use strict'
 
-const constants = require('./constants')
 const errcode = require('err-code')
+const constants = require('./constants')
+const getGossipPeers = require('./getGossipPeers')
 const { shuffle } = require('./utils')
 
 class Heartbeat {
@@ -77,7 +78,7 @@ class Heartbeat {
       // do we have enough peers?
       if (peers.size < constants.GossipSubDlo) {
         const ineed = constants.GossipSubD - peers.size
-        const peersSet = this.gossipsub._getGossipPeers(topic, ineed)
+        const peersSet = getGossipPeers(this.gossipsub, topic, ineed)
         peersSet.forEach((peer) => {
           // add topic peers not already in mesh
           if (peers.has(peer)) {
@@ -138,7 +139,7 @@ class Heartbeat {
       // do we need more peers?
       if (peers.size < constants.GossipSubD) {
         const ineed = constants.GossipSubD - peers.size
-        const peersSet = this.gossipsub._getGossipPeers(topic, ineed)
+        const peersSet = getGossipPeers(this.gossipsub, topic, ineed)
         peersSet.forEach((peer) => {
           if (!peers.has(peer)) {
             return
